refactor: Pentagram - SelectView, UpdateView 컴포넌트 리팩토링
